+++ packages/eyes-universal/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
 ## Unreleased
+
+
+## 1.0.3 - 2022/1/17
+
 - updated to @applitools/eyes-sdk-core@12.24.14 (from 12.24.10)
 - updated to @applitools/visual-grid-client@15.8.61 (from 15.8.56)
-
 
 ## 1.0.2 - 2022/1/6